Initial commit
Libraries: contains external dependencies like the u8glib library that is necessary for the GLCD Display
MarlinSingle: Contains a ready to go firmware for the single extruder MendelMax 3 setup - nearly vanilla as provided by Maker's Tool Works
MarlinDual: This is a modified version of the MendelMax 3 Marlin firmware with support for the dual extruder setup. Additionally I removed the third extruder icon and switched the extruder icons (and value prints) so that they match with the physical extruder setup where the primary extruder is on the right and the secondary on the left side of the assembly.